refactor: post module package cleanup
