+++ gateway/src/main/java/com/oing/controller/CalendarController.java
@@ -2,7 +2,7 @@ package com.oing.controller;
 
 import com.oing.component.TokenAuthenticationHolder;
 import com.oing.domain.MemberPostCountDTO;
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import com.oing.dto.response.ArrayResponse;
 import com.oing.dto.response.CalendarResponse;
 import com.oing.restapi.CalendarApi;
