+++ gateway/src/main/java/com/oing/controller/WidgetController.java
@@ -3,7 +3,7 @@ package com.oing.controller;
 
 import com.oing.component.TokenAuthenticationHolder;
 import com.oing.domain.Member;
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import com.oing.dto.response.SingleRecentPostWidgetResponse;
 import com.oing.restapi.WidgetApi;
 import com.oing.service.MemberPostService;
