+++ post/src/main/java/com/oing/controller/PostController.java
@@ -2,7 +2,7 @@ package com.oing.controller;
 
 
 import com.oing.domain.PaginationDTO;
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import com.oing.dto.request.CreatePostRequest;
 import com.oing.dto.request.PreSignedUrlRequest;
 import com.oing.dto.response.PaginationResponse;
