+++ post/src/main/java/com/oing/domain/MemberPost.java
@@ -1,6 +1,5 @@
-package com.oing.domain.model;
+package com.oing.domain;
 
-import com.oing.domain.BaseAuditEntity;
 import jakarta.persistence.*;
 import lombok.*;
