+++ post/src/main/java/com/oing/domain/MemberPostReaction.java
@@ -1,7 +1,5 @@
-package com.oing.domain.model;
+package com.oing.domain;
 
-import com.oing.domain.BaseEntity;
-import com.oing.domain.Emoji;
 import jakarta.persistence.*;
 import lombok.*;
