+++ post/src/main/java/com/oing/dto/response/PostResponse.java
@@ -1,6 +1,6 @@
 package com.oing.dto.response;
 
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.ZoneId;
