+++ post/src/main/java/com/oing/repository/MemberPostCommentRepository.java
@@ -1,6 +1,6 @@
 package com.oing.repository;
 
-import com.oing.domain.model.MemberPostComment;
+import com.oing.domain.MemberPostComment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberPostCommentRepository extends JpaRepository<MemberPostComment, String> {
