+++ post/src/main/java/com/oing/repository/MemberPostRepository.java
@@ -1,6 +1,6 @@
 package com.oing.repository;
 
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
