+++ post/src/main/java/com/oing/service/MemberPostService.java
@@ -4,7 +4,7 @@ import com.oing.domain.MemberPostCountDTO;
 import com.oing.domain.PaginationDTO;
 import com.oing.exception.DomainException;
 import com.oing.exception.ErrorCode;
-import com.oing.domain.model.MemberPost;
+import com.oing.domain.MemberPost;
 import com.oing.exception.PostNotFoundException;
 import com.oing.repository.MemberPostRepository;
 import com.querydsl.core.QueryResults;
