+++ post/src/test/java/com/oing/domain/model/MemberPostCommentTest.java
@@ -1,5 +1,7 @@
 package com.oing.domain.model;
 
+import com.oing.domain.MemberPost;
+import com.oing.domain.MemberPostComment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
