+++ post/src/test/java/com/oing/domain/model/MemberPostReactionTest.java
@@ -1,6 +1,8 @@
 package com.oing.domain.model;
 
 import com.oing.domain.Emoji;
+import com.oing.domain.MemberPost;
+import com.oing.domain.MemberPostReaction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
